Support second/minute/hour suffixes when prompting for number of seconds

+++ testmocking/management/commands/answer_all_tasks.py
@@ -21,7 +21,7 @@ class Command(BaseCommand):
         parser.add_argument('--org', type=str, required=True, help="the slug for the organization (group) to target")
         parser.add_argument('--impute', choices=['halt', 'skip', 'answer'], default='halt', help="specify 'halt' (the default) to abort on the first non-handled imputed question, 'skip' to ignore it and answer future questions, or 'answer' to answer it despite possibly being imputed")
         parser.add_argument('--quiet', action='store_true', help="Reduces verbosity")
-        parser.add_argument('--delay', default=0, type=int, help="Number of seconds to wait between each action. Must be non-negative (0 is no delay)")
+        parser.add_argument('--delay', default=0, type=float, help="Number of seconds to wait between each action. Must be non-negative (0 is no delay)")
 
     def handle(self, *args, **options):
         def delay():
